% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkjustgage.R
\name{dkjustgage}
\alias{dkjustgage}
\title{HTMLWidget wrapper for justgage}
\usage{
dkjustgage(value = 5, min = 0, max = 100, title = "", label = "",
  reverse = F, symbol = "", target = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{value}{Gage value}

\item{min, max}{Gage limits}

\item{title}{Title above gage}

\item{label}{Text label below the gage eg units}

\item{reverse}{True to reverse color gradient}

\item{symbol}{Postfix a symbol to the value label eg \code{'\%'}}

\item{target}{Set a traffice light target value to acheive, red if below, green if above}
}
\description{
Radial gage widget
}
\examples{
dkjustgage(50,0,100,title="speed")
}

